Optimizations to tally_scoring
As nuc_xs now selectively points to nuc_xs or macro_xs, most of the `if
(nuclide >= 0) { } else {} can actually removed.
Need to come up with a better name for nuc_xs.
Some if/else are still there but may be removed as well. But I can't
judge, if this introduces some side effects. (e.g. when
p.macro_xs().total is used instead of macro_xs(MgxsType::TOTAL)